// src/app/models/task.model.ts
export interface Task {

    key: string;
    id: string;
    name: string;
    description: string;
    timeOfCreation: string;

    // constructor(id: number, name: string, description: string, timeOfCreation: string) {

    //     this.id = id;
    //     this.name = name;
    //     this.description = description;
    //     this.timeOfCreation = timeOfCreation;
    // }
}